+++ tests/test_regex_matcher.py
@@ -786,6 +786,10 @@ def test_split():
     assert dest1[0] == "foo"
     assert dest1[1] == "bar baz"
 
+    with pytest.raises(ICUError) as exc_info:
+        _ = matcher.split(src1, dest1, 0)
+    assert exc_info.value.args[0] == UErrorCode.U_ILLEGAL_ARGUMENT_ERROR
+
     # [2]
     # int32_t icu::RegexMatcher::split(
     #       UText *input,
@@ -794,12 +798,16 @@ def test_split():
     #       UErrorCode &status
     # )
     src2 = utext_open_const_unicode_string(None, src1)
-    dest2 = UTextVector(10)
+    out = UnicodeStringVector(10)
+    dest2 = UTextVector(out)
     result = matcher.split(src2, dest2)
     assert result == 3
     assert utext_extract(dest2[0], 0, utext_native_length(dest2[0])) == "foo"
     assert utext_extract(dest2[1], 0, utext_native_length(dest2[1])) == "bar"
     assert utext_extract(dest2[2], 0, utext_native_length(dest2[2])) == "baz"
+    assert out[0] == "foo"
+    assert out[1] == "bar"
+    assert out[2] == "baz"
 
     result = matcher.split(src2, dest2, 2)
     assert result == 2
@@ -807,6 +815,13 @@ def test_split():
     assert (
         utext_extract(dest2[1], 0, utext_native_length(dest2[1])) == "bar baz"
     )
+    assert out[0] == "foo"
+    assert out[1] == "bar baz"
+
+    with pytest.raises(ICUError) as exc_info:
+        _ = matcher.split(src2, dest2, 0)
+    assert exc_info.value.args[0] == UErrorCode.U_ILLEGAL_ARGUMENT_ERROR
+
     utext_close(src2)